third_party/blink/web_tests/fast/forms/editing-value-expected.txt

This tests setting the editing value of an input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS oninput event was fired.
PASS input.value is "foo"
PASS document.querySelector(":invalid") is input
PASS onchange event was fired.
PASS successfullyParsed is true

TEST COMPLETE
